pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "bookings-etl"
version = "0.1.0"
description = "A Dagster project for booking ETL."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
dagster = "^1.0.0"
dagit = "^1.0.0"
python-dotenv = "^0.19.2"
supabase = "^0.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[tool.dagster]
module = "bookings_etl"
